fix: timestamp conversion for 32bit platforms

+++ pkg/analyze/dir_linux-openbsd.go
@@ -1,3 +1,5 @@
+// +build linux openbsd
+
 package analyze
 
 import (
@@ -12,7 +14,7 @@ func setPlatformSpecificAttrs(file *File, f os.FileInfo) {
 	switch stat := f.Sys().(type) {
 	case *syscall.Stat_t:
 		file.Usage = stat.Blocks * devBSize
-		file.Mtime = time.Unix(stat.Mtim.Sec, stat.Mtim.Nsec)
+		file.Mtime = time.Unix(int64(stat.Mtim.Sec), int64(stat.Mtim.Nsec))
 
 		if stat.Nlink > 1 {
 			file.Mli = stat.Ino
@@ -26,5 +28,5 @@ func setDirPlatformSpecificAttrs(dir *Dir, path string) {
 		return
 	}
 
-	dir.Mtime = time.Unix(stat.Mtim.Sec, stat.Mtim.Nsec)
+	dir.Mtime = time.Unix(int64(stat.Mtim.Sec), int64(stat.Mtim.Nsec))
 }
